Eliminate (most) "signed and unsigned comparison" warnings from library code [-Wsign-compare]
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@19017 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/check_varbind.c
@@ -58,7 +58,7 @@ netsnmp_check_vb_range(const netsnmp_variable_list *var,
     if (NULL == var)
         return SNMP_ERR_GENERR;
     
-    if ((*var->val.integer < low) || (*var->val.integer > high)) {
+    if (((size_t)*var->val.integer < low) || ((size_t)*var->val.integer > high)) {
         rc = SNMP_ERR_WRONGVALUE;
     }
